requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.3.1
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
attrs=25.3.0=pypi_0
blinker=1.9.0=pypi_0
brotli=1.1.0=pypi_0
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
cssselect2=0.8.0=pypi_0
et-xmlfile=2.0.0=pypi_0
exceptiongroup=1.3.0=pypi_0
fonttools=4.58.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
idna=3.10=pypi_0
iniconfig=2.1.0=pypi_0
jinja2=3.1.6=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
narwhals=1.40.0=pypi_0
numpy=2.0.2=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.16=h3f729d1_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_2
plotly=6.1.1=pypi_0
pluggy=1.6.0=pypi_0
protobuf=6.31.0=pypi_0
pyarrow=20.0.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pydeck=0.9.1=pypi_0
pydyf=0.11.0=pypi_0
pygments=2.19.1=pypi_0
pypdf2=3.0.1=pypi_0
pyphen=0.17.2=pypi_0
pytest=8.3.5=pypi_0
python=3.9.21=h8205438_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
referencing=0.36.2=pypi_0
requests=2.32.3=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.25.1=pypi_0
setuptools=78.1.1=py39haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sqlite=3.45.3=h2bbff1b_0
streamlit=1.45.1=pypi_0
tenacity=9.1.2=pypi_0
tinycss2=1.4.0=pypi_0
tinyhtml5=2.0.0=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
tornado=6.5=pypi_0
tqdm=4.67.1=pypi_0
typer=0.15.4=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
watchdog=6.0.0=pypi_0
weasyprint=65.1=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.45.1=py39haa95532_0
zopfli=0.2.3.post1=pypi_0
